requirements.txt
pysimplegui
cryptodomex